Reorganized api.temp into its own module so that we can organize routes by functionality
